fix test and rename
