+++ straxen/storage/__init__.py
@@ -1,11 +1,11 @@
 from . import online_monitor
 from .online_monitor import *
 
-from . import rucio
-from .rucio import *
+from . import rucio_remote
+from .rucio_remote import *
 
-from .local_rucio import *
-from . import local_rucio
+from .rucio_local import *
+from . import rucio_local
 
 from . import rundb
 from .rundb import *
